+++ src/index.ts
@@ -1,4 +1,4 @@
 export * from './mini-signals.ts';
-export * from './mini-signals-bus.ts';
+export * from './mini-signals-emitter.ts';
 export * from './mini-signals-utils.ts';
 export type * from './types.d.ts';