+++ src/mini-signals-utils.ts
@@ -1,15 +1,15 @@
-import { MiniSignal } from "./mini-signals.ts";
-import type { MiniSignalEventMap } from "./types.d.ts";
+import { MiniSignal } from './mini-signals.ts';
+import type { EventMap, SignalMap } from './types.d.ts';
 
 /**
  * Helper to create a signal map for SignalEmitter
  */
-export function createSignalMap<T extends MiniSignalEventMap>(
-  events: (keyof T)[]
-): { [K in keyof T]: MiniSignal<T[K]> } {
-  const map = {} as { [K in keyof T]: MiniSignal<T[K]> };
+export function createSignalMap<T extends EventMap>(
+  events: Array<keyof T>
+): SignalMap<T> {
+  const map: Partial<SignalMap<T>> = {};
   for (const event of events) {
     map[event] = new MiniSignal<T[typeof event]>();
   }
-  return map;
+  return map as SignalMap<T>;
 }