+++ src/mini-signals.test-d.ts
@@ -1,6 +1,8 @@
-import { expectAssignable, expectError, expectType } from 'tsd';
+import { expectError, expectType } from 'tsd';
 import { describe, it } from 'vitest';
+
 import { MiniSignal } from './mini-signals.ts';
+import { MiniSignalEmitter } from './mini-signals-emitter.ts';
 
 describe('MiniSignal Typing', () => {
   it('should have correct types', () => {
@@ -70,5 +72,71 @@ describe('MiniSignal Typing', () => {
     expectError(e1.detach(l2));
 
     expectType<boolean>(e1.dispatch('foo'));
+  });
+});
+
+describe('MiniSignalEmitter Typing', () => {
+  it('should have correct types', () => {
+    const emitter = new MiniSignalEmitter({
+      login: new MiniSignal<[string]>(),
+      logout: new MiniSignal<[string]>(),
+    });
+
+    const offLogin = emitter.on('login', (userId) => {
+      expectType<string>(userId);
+    });
+
+    const offLogout = emitter.once('logout', (reason) => {
+      expectType<string>(reason);
+    });
+    expectType<boolean>(emitter.emit('login', 'user123'));
+    expectType<boolean>(emitter.emit('logout', 'timeout'));
+    offLogin();
+    offLogout();
+  });
+
+  it('should have correct types (symbols)', () => {
+    const LOGIN = Symbol('login');
+    const LOGOUT = Symbol('logout');
+
+    const emitter = new MiniSignalEmitter({
+      [LOGIN]: new MiniSignal<[string]>(),
+      [LOGOUT]: new MiniSignal<[string]>(),
+    });
+
+    const offLogin = emitter.on(LOGIN, (userId) => {
+      expectType<string>(userId);
+    });
+
+    const offLogout = emitter.once(LOGOUT, (reason) => {
+      expectType<string>(reason);
+    });
+    expectType<boolean>(emitter.emit(LOGIN, 'user123'));
+    expectType<boolean>(emitter.emit(LOGOUT, 'timeout'));
+
+    offLogin();
+    offLogout();
+  });
+
+  it('should show TS error on incorrect listeners and dispatch', () => {
+    const emitter = new MiniSignalEmitter({
+      login: new MiniSignal<[string]>(),
+      logout: new MiniSignal<[string]>(),
+    });
+
+    expectError(
+      emitter.on('login', (userId: number) => {
+        expectType<string>(userId);
+      })
+    );
+
+    expectError(
+      emitter.once('logouts', (reason) => {
+        expectType<string>(reason);
+      })
+    );
+
+    expectError(emitter.emit('login', 123));
+    expectError(emitter.emit('logouts', 'timeout'));
   });
 });