+++ CHANGELOG.md
@@ -1,13 +1,13 @@
 # CHANGELOG
 
 ## HEAD (Unreleased)
-_(none)_
+* Breaking: possible breaking change on imports, now exports `index.cjs` and `index.mjs`
+* Feature: Added `MiniSignalEmitter`
 
 ---
 
 ## 2.1.0 (2026-01-01)
-* Adding `.dispatchSerial`
-* Adding `.dispatchParallel`
+* Feature: adding `.dispatchSerial` and `.dispatchParallel`
 
 ## 2.0.0 (2023-03-29)